Clean up PlatformAgnostic SystemInfo
Primarily motivated by removing ChakraCore's dependency on
the GetPhysicalMemory Win32 API on Windows. This is problematic
from a UWP compliance perspective, so removed usage of that API
altogether.

Also did some cleanup- the existing code had ch have a hard dependency
on code in lib/Runtime/PlatformAgnostic- changed this to follow the existing
pattern of using ChakraRTInterface as an abstraction.

Also made some stylistic changes to make the code more consistent with
the existing style in the file.